Scrounger checks nearby containers
For some reason I decided just to --comment Scrounger trait and its Event for now.

Newly added function c replaces Scrounger and does everything Scrounger did, but now it scans nearby containers instead of scanning them on player loading a chunk or however it did that in the past. It then tripple checks if it is a valid container and if its inventory was already altered (if not containerObj:getModData().bScroungerRolled), (not nil is true).
Only then roll is made as in old Scrounger function.
Additional prints help track what item was added and in what quantity, they can be removed later ofc.