added compatibility for indefatigable and injured
the way it works, it adds the affected limbs to the table. those limbs will not heal if indefatigable triggers, this also means that any wounds caused by something else WILL stay there. game checks every hour if a limb from the table has any injuries, and if it doesn't, it gets removed from the table, allowing indefatigable to heal it.