Gordanite Bloodied Crowbar Fix
Added test to "old gordanite" check on line 4099 to compare TreeDamage to listed damage of a Bloodied Crowbar, which is set to 100 by default. This prevents old gordanite code from running against a Bloodied Crowbar (just like the >0 test prevents it running against a vanilla Crowbar) and thus prevents it from breaking the stats or name string of the Bloodied Crowbar when Gordanite is present.